Add files via upload
# Plagiarism Checker

## Overview

This Plagiarism Checker project is a simple tool to assess the similarity between two texts using the difflib library. It provides a basic interface using Tkinter for inputting the text to be checked and the reference text.

## Author

*Author:* YADAM VIJAY KUMAR

*LinkedIn:* [YADAM VIJAY KUMAR] (https://www.linkedin.com/in/vijaykumar-yadam-300155264)

## Project Structure

The project is structured as a Python script utilizing difflib for text comparison and Tkinter for the graphical user interface. Key features include:

- Text similarity calculation
- Plagiarism detection with a user-friendly interface

## Usage

1. Run the script.
2. Enter your text in the input text widget.
3. Enter the reference text in the reference text widget.
4. Click the "Check Plagiarism" button.
5. The result will be displayed in the result text widget.

## Dependencies

No additional dependencies are required for this project.

## Contribution

Feel free to contribute or provide feedback to enhance the functionality of this simple Plagiarism Checker. Your suggestions are valuable!

## About the Author

Pydisetti Vineesh is currently interning under Codetice Company. Connect with him on [LinkedIn](https://www.linkedin.com/in/vineesh-pydisetti-28464421b/) for more updates and discussions.

Happy checking for plagiarism! 📚✨